refactor: move mount/env spec builders to api/v1/spec_utils
make_mounts_from_strings / make_env_vars_from_strings were parked in
api/v1/deployment.py during the photon removal, but they are generic
deployment-spec builders shared by the endpoint, pod, job, finetune, and
raycluster commands. Move them to a dedicated api/v1/spec_utils.py so the
DeploymentAPI module isn't their home. The 5 CLI importers are repointed.

make_token_vars_from_config stays in deployment.py (it predates this work
and is deployment-token-specific).
