+++ leptonai/cli/pod.py
@@ -39,7 +39,7 @@ from .util import (
 )
 from .util import make_container_ports_from_str_list
 from ..api.v2.client import APIClient
-from ..api.v1.deployment import make_mounts_from_strings, make_env_vars_from_strings
+from ..api.v1.spec_utils import make_mounts_from_strings, make_env_vars_from_strings
 from ..api.v1.types.deployment import (
     LeptonDeploymentState,
     ResourceRequirement,
